Slight reorganization.
Placed a new iterator "in front of" the previous lexer code.  This is now
the interface to the world.  It serves the purpose of merging multiple
consecutive string tokens to support string literal concatenation.

Adjacent string literals (thus far only the simplest version) are now
automatically concatenated into a single token even if on separate physical
lines but on the same logical line.  At present, only explicit line joins
are supported, but this same concatenation technique should work as well
with implicit line joins because it works on the token stream.