Split the error checking of escaped characters into two phases.  The first simply checks for a structurally invalid escape sequence.  The second phase (if the first passes) checks for invalid unicode character names by iterating over all structurally valid named escape sequences.
This second phase is less efficient than I would like because, if it passes,
the substitution phase must also iterate over all of these sequences.  I
can improve this by checking while substituting, but the tangle of concerns
leads to unpleasant code.  If this is ever an issue (ha?), then it can be
changed.